+++ src/norm/private/postgres/dbtypes.nim
@@ -9,25 +9,36 @@ To add support for ``YourType``, define three funcs:
 
 import options
 import times
+import strutils
 
 import ndb/postgres
 
-import ../../model
+import ../../model, ../../types
 
 
 # Funcs that return an SQLite type for a given Nim type:
 
-func dbType*(T: typedesc[SomeInteger]): string = "INTEGER"
+func dbType*(T: typedesc[int16]): string = "SMALLINT"
 
-func dbType*(T: typedesc[SomeFloat]): string = "REAL"
+func dbType*(T: typedesc[int32]): string = "INTEGER"
+
+func dbType*(T: typedesc[int64]): string = "BIGINT"
+
+func dbType*(T: typedesc[float32]): string = "REAL"
+
+func dbType*(T: typedesc[float64]): string = "DOUBLE PRECISION"
 
 func dbType*(T: typedesc[string]): string = "TEXT"
+
+func dbType*[C](_: typedesc[StringOfCap[C]]): string = "VARCHAR($#)" % $C
+
+func dbType*[C](_: typedesc[PaddedStringOfCap[C]]): string = "CHAR($#)" % $C
 
 func dbType*(T: typedesc[bool]): string = "BOOLEAN"
 
 func dbType*(T: typedesc[DateTime]): string = "TIMESTAMP WITH TIME ZONE"
 
-func dbType*(T: typedesc[Model]): string = "INTEGER"
+func dbType*(T: typedesc[Model]): string = "BIGINT"
 
 func dbType*[T](_: typedesc[Option[T]]): string = dbType T
 
@@ -42,6 +53,10 @@ func dbValue*[T: Model](val: Option[T]): DbValue =
   else:
     dbValue(nil)
 
+func dbValue*[_](val: StringOfCap[_]): DbValue = dbValue(string(val))
+
+func dbValue*[_](val: PaddedStringOfCap[_]): DbValue = dbValue(string(val))
+
 
 # Converter funcs from ``DbValue`` instances to Nim types:
 
@@ -52,6 +67,10 @@ func to*(dbVal; T: typedesc[SomeInteger]): T = dbVal.i.T
 func to*(dbVal; T: typedesc[SomeFloat]): T = dbVal.f.T
 
 func to*(dbVal; T: typedesc[string]): T = dbVal.s
+
+func to*[_](dbVal; T: typedesc[StringOfCap[_]]): T = dbVal.o.value.T
+
+func to*[_](dbVal; T: typedesc[PaddedStringOfCap[_]]): T = dbVal.o.value.T
 
 func to*(dbVal; T: typedesc[bool]): T = dbVal.b